Further extract lint/format base
