+++ src/BaseFormatter.php
@@ -5,31 +5,11 @@ namespace Tighten\TLint;
 use PhpParser\Lexer;
 use PhpParser\Parser;
 
-class BaseFormatter
+class BaseFormatter extends Base
 {
-    public const DESCRIPTION = 'No Description for Formatter.';
-
-    protected $description;
-    protected $filename;
-    protected $code;
-    protected $codeLines;
-
-    public function __construct($code, $filename = null)
+    public function format(Parser $parser, Lexer $lexer): string
     {
-        $this->description = static::DESCRIPTION;
-        $this->filename = $filename;
-        $this->code = $code;
-        $this->codeLines = preg_split('/\r\n|\r|\n/', $code);
-    }
-
-    public static function appliesToPath(string $path, array $configPaths): bool
-    {
-        return true;
-    }
-
-    public function format(Parser $parser, Lexer $lexer)
-    {
-        return [];
+        return $this->code;
     }
 
     public function getFormatDescription()
@@ -40,15 +20,5 @@ class BaseFormatter
     public function setFormatDescription(string $description)
     {
         return $this->description = $description;
-    }
-
-    public function getCode(): string
-    {
-        return $this->code;
-    }
-
-    public function getFilename()
-    {
-        return $this->filename;
     }
 }
